api/powerplant/python-document-api/requirements.txt
Flask
python-docx
reportlab
Jinja2
PyPDF2